refactor(agents): playground-kit barrel, project-fs core, skills hot-reload fix, dead-code delete
PR 2 of docs/design/playground.md §13:
- playground/index.ts (@aep/agents/playground-kit): curated re-exports of the
  reusable in-package playground pieces (project-fs, render, derived, mcp).
- playground/project-fs.ts: the dir-parameterized read/reconcile/fence core
  extracted from threads.ts; threads.ts keeps naming+listing and delegates.
- EvalWorkspace.materializeSkills no longer memoizes the first library per
  instance — content-addressed per call, so an edited SKILL.md yields a new
  snapshot on the next turn (hot-reload); regression tests added.
- DELETED dead playground/tasks.ts + evals/taskplanner/ (imported the removed
  src/agents/taskplanner module; typecheck:playground was red because of it).
- fixed pre-existing playground-lane breakage: mcp.ts imported TurnRequest
  from the deleted src/contracts/sse-events.js (now @aep/agent-stream), and
  dsl.test.ts still used the retired coordinate dialect fixture.

All three lanes green: test 147, test:eval 42, test:playground 17; lint +
typecheck/typecheck:eval/typecheck:playground clean.
