--- services/agents/evals/taskplanner/fixtures/all-four-kinds.json
@@ -1,53 +1,0 @@
-{
-  "name": "all-four-kinds",
-  "difficulty": "medium",
-  "input": {
-    "projectName": "orders-platform",
-    "mode": "fresh",
-    "spec": "# Orders platform\n\nA storefront where a customer browses products, places an order, and pays. The web app talks to the orders service; the orders service records orders in its database and charges the customer through Stripe; it identifies the customer via the org-wide identity service.",
-    "slimDesign": [
-      {
-        "name": "payments-api",
-        "componentType": "service",
-        "language": "Go",
-        "dependsOn": [],
-        "externalResources": [
-          { "name": "stripe", "description": "Stripe payment gateway — API key + webhook secret" }
-        ],
-        "platformResources": [
-          { "name": "payments-db", "resourceType": "postgres-cnpg", "description": "orders/payments ledger" }
-        ]
-      },
-      {
-        "name": "orders-api",
-        "componentType": "service",
-        "language": "Go",
-        "dependsOn": ["payments-api"],
-        "orgServiceDependencies": [
-          { "name": "identity", "description": "org-wide identity/JWT service (another project)" }
-        ]
-      },
-      {
-        "name": "orders-web",
-        "componentType": "web-application",
-        "language": "TypeScript",
-        "dependsOn": ["orders-api"]
-      }
-    ],
-    "attachedSkills": [
-      { "name": "go", "description": "Go service stack conventions" },
-      { "name": "react-webapp", "description": "React SPA conventions" }
-    ],
-    "diff": { "added": ["payments-api", "orders-api", "orders-web"], "contractAffectedModified": [], "removed": [] }
-  },
-  "expect": {
-    "coverAllComponents": true,
-    "providerBeforeConsumer": [
-      ["orders-api", "payments-api"],
-      ["orders-web", "orders-api"]
-    ],
-    "gatesInRationale": {
-      "payments-api": ["stripe", "payments-db"]
-    }
-  }
-}
